Alertas para crud de clientes y productos(para productos hacen falta los botones guardar modificacion, eliminar). Sigue vigente lo del commit anterior: -> Solucionado el error cannot get /, se eliminó [ngModelOptions]="{standalone: true}" de la linea 61 del archivo clientes/component/visitasmodal/visitasmodal.component.html

+++ SegundoDesafio/veterinariaDamsi/src/app/productos/producto/producto.component.ts
@@ -8,6 +8,8 @@ import { ProductoService } from '../../services/producto.service';
 //Validacion de formularios
 import { FormGroup, FormControl, Validators } from '@angular/forms';
 
+import Swal from 'sweetalert2';
+
 
 @Component({
   selector: 'app-producto',
@@ -33,18 +35,82 @@ export class ProductoComponent implements OnInit {
     //Añade el producto a la bd
     onSubmit(productoForm:NgForm){ 
         let producto= this.productoServicio.productoSeleccionado;
-        this.productoServicio.crearProducto(producto);        
+        
+        Swal.fire({
+          title: '¿Agregar un nuevo producto?',
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#606ED7',
+          confirmButtonText: 'Sí, Agregar'
+        }).then(result => {
+          if(result.value){
+            Swal.fire({
+              title: 'Cargando...',
+              didOpen(){
+                Swal.showLoading();
+              }
+              //si
+            })//.then(result => {
+              this.productoServicio.crearProducto(producto).then(result =>{
+            Swal.fire('¡Producto Guarado!', 'Se ha guardado la información del producto', 'success')
+            //})
+          }).catch((error) => {
+              Swal.fire({
+                title: 'Error!',
+                text: 'No se pudo guardar',
+                icon:'error'
+              })
+            })
+          }
+          else{
+            Swal.fire({
+              position: 'top-end',
+              title: 'Acción Cancelada...',
+              text: `Acción crear nuevo producto cancelada`,
+              icon: 'info',
+              heightAuto:true,
+              timer: 2000,
+              timerProgressBar: true
+          });
+          }
+          });
+        //this.productoServicio.crearProducto(producto);        
     }
    
      //actualiza la informacion del producto
     actualizar(){ 
-      this.productoServicio.actualizarProducto();
+      Swal.fire({
+        title: 'Cargando...',
+        didOpen(){
+          Swal.showLoading()
+        },
+        didClose (){
+          Swal.hideLoading()
+        }
+      });
+      this.productoServicio.actualizarProducto().then(result =>{
+        Swal.fire('¡Información actualizada!', 'Los datos se han actualizado con exito', 'success');
+      });
+      
       this.productoServicio.cancelarSeleccion(); 
     }
 
     //Envia el cliente para su eliminacion
     eliminar(){ 
-      this.productoServicio.eliminarProducto();
+      Swal.fire({
+        title: '¿Está seguro que desea eliminar el producto seleccionado?',
+        text: `¡Esta acción no se puede deshacer!`,
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#606ED7',
+        confirmButtonText: 'Sí, eliminar'
+      }).then(result => {
+        if(result.value){
+          this.productoServicio.eliminarProducto(); 
+          Swal.fire('¡Producto Elimiando!', 'El producto se ha eliminado con exito', 'success');
+        }
+      }) 
+      
     }
 
     @HostListener('document:keydown', ['$event'])
